NTBS-2832: Create share pages, show message when case shared
