+++ ntbs-service/Models/Entities/HospitalDetails.cs
@@ -4,6 +4,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using EFAuditer;
 using ExpressiveAnnotations.Attributes;
+using Newtonsoft.Json.Serialization;
 using ntbs_service.Models.ReferenceEntities;
 using ntbs_service.Models.Validations;
 
@@ -36,6 +37,7 @@ namespace ntbs_service.Models.Entities
         public virtual Hospital Hospital { get; set; }
 
         [Display(Name = "TB Service record is shared with")]
+        [AssertThat("SecondaryTBServiceCode != TBServiceCode", ErrorMessage = ValidationMessages.ShareDestinationCannotBeCurrentTbService)]
         public string SecondaryTBServiceCode { get; set; }
         public virtual TBService SecondaryTBService { get; set; }
