--- ntbs-service/Pages/Alerts/Dismiss.cshtml
@@ -1,7 +1,0 @@
-@page "/Alerts/{AlertId:int}/Dismiss/"
-@model ntbs_service.Pages.Alerts.DismissModel
-
-@{
-    ViewData["Title"] = "Dismiss Notification";
-}
-<!-- This page redirects on GET, so this should never be displayed. -->
