--- ntbs-service/Pages/Alerts/_AcceptedTransferConfirmation.cshtml
@@ -1,17 +1,0 @@
-@using ntbs_service.Helpers
-@model ntbs_service.Pages.Alerts.TransferRequestActionModel
-
-@{
-    Layout = "../Shared/_NotificationLayout.cshtml";
-    ViewData["Title"] = $"Transfer Request Accepted - {Model.NotificationId}";
-}
-
-<nhs-width-container container-width="Standard">
-    <h2> Notification #@Model.NotificationId has been accepted </h2>
-    <div> You can now edit this notification </div>
-    <br>
-    <a id="return-to-notification" href="@RouteHelper.GetNotificationPath(Model.NotificationId, NotificationSubPaths.Overview)" 
-        class="nhsuk-button ntbsuk-button--primary return-to-homepage-link">
-        Return to the notification
-    </a>
-</nhs-width-container>
