--- ntbs-service/Pages/Alerts/_RejectedTransferConfirmation.cshtml
@@ -1,16 +1,0 @@
-@using ntbs_service.Helpers
-@model ntbs_service.Pages.Alerts.TransferRequestActionModel
-
-@{
-    Layout = "../Shared/_NotificationLayout.cshtml";
-    ViewData["Title"] = $"Transfer Request Rejected - {Model.NotificationId}";
-}
-
-<nhs-width-container container-width="Standard">
-    <h2> Transfer of Notification #@Model.NotificationId has been rejected </h2>
-    <br>
-    <a id="return-to-homepage" href="@RouteHelper.GetNotificationPath(Model.NotificationId, NotificationSubPaths.Overview)" 
-        class="nhsuk-button ntbsuk-button--primary return-to-homepage-link">
-        Return to the notification
-    </a>
-</nhs-width-container>
