--- ntbs-service/Pages/Alerts/_TransferPendingPartial.cshtml
@@ -1,91 +1,0 @@
-@using ntbs_service.Helpers
-@model ntbs_service.Pages.Alerts.TransferRequestModel
-
-@{
-    Layout = "../Shared/_NotificationLayout.cshtml";
-    ViewData["Title"] = $"Transfer Request Pending - {Model.NotificationId}";
-}
-
-<nhs-width-container container-width="Standard">
-    <form method="post" autocomplete="off">
-        <input type="hidden" asp-for="TransferAlert.NotificationId" value="@Model.NotificationId" />
-        <h3> Notification #@Model.NotificationId is pending transfer acceptance </h3>
-
-        <div>
-            It will continue to appear in your Notifications and you can continue to edit until it is accepted.
-        </div>
-
-        <br>
-
-        <h3> Transfer summary </h3>
-
-        <div class="bold-text"> Transfer sent to: </div>
-        <nhs-grid-row>
-            <nhs-grid-column grid-column-width="OneThird">
-                <span asp-for="TransferAlert.TbServiceCode" class="nhsuk-label bold-text">
-                    Region
-                </span>
-                <div id="transfer-region">
-                    @Model.TransferAlert.TbService?.PHEC?.Name
-                </div>
-            </nhs-grid-column>
-
-            <nhs-grid-column grid-column-width="OneThird">
-                <span asp-for="TransferAlert.TbServiceCode" class="nhsuk-label bold-text">
-                    TB Service
-                </span>
-                <div id="transfer-tb-service">
-                    @Model.TransferAlert.TbService?.Name
-                </div>
-            </nhs-grid-column>
-
-            <nhs-grid-column grid-column-width="OneThird">
-                <span asp-for="TransferAlert.TransferDate" class="nhsuk-label bold-text">
-                    Transfer date
-                </span>
-                <div id="transfer-date">
-                    @Model.TransferAlert.TransferDate?.ConvertToString()
-                </div>
-            </nhs-grid-column>
-
-            <nhs-grid-column grid-column-width="OneThird">
-                <span asp-for="TransferAlert.TbServiceCode" class="nhsuk-label bold-text">
-                    Case manager
-                </span>
-                <div id="transfer-case-manager">
-                    @Model.TransferAlert.CaseManager?.DisplayName
-                </div>
-            </nhs-grid-column>
-        </nhs-grid-row>
-
-        <br>
-
-        <span asp-for="TransferAlert.TbServiceCode" class="nhsuk-label bold-text">
-            Reason for transfer
-        </span>
-        <div>
-            @Model.TransferAlert.TransferReasonString
-        </div>
-
-        <br>
-
-        <span asp-for="TransferAlert.TbServiceCode" class="nhsuk-label bold-text">
-            Additional information sent with transfer
-        </span>
-        <div>
-            @Model.TransferAlert.TransferRequestNote
-        </div>
-
-        <br>
-
-        <div class="flex-container">
-            <a href="@RouteHelper.GetNotificationPath(Model.NotificationId, NotificationSubPaths.Overview)"
-                class="nhsuk-button ntbsuk-button--primary pending-transfer-return-link">
-                Return to notification
-            </a>
-            <button id="cancel-transfer-button" asp-page-handler="Cancel" class="button-link pending-transfer-cancel-button">
-                Cancel transfer
-            </button>
-        </div>
-    </form>
-</nhs-width-container>
