+++ ntbs-service/Pages/AlertsAndActions/Dismiss.cshtml
@@ -1,0 +1,7 @@
+@page "/Alerts/{AlertId:int}/Dismiss/"
+@model ntbs_service.Pages.AlertsAndActions.DismissModel
+
+@{
+    ViewData["Title"] = "Dismiss Notification";
+}
+<!-- This page redirects on GET, so this should never be displayed. -->
