+++ ntbs-service/Pages/AlertsAndActions/ShareWithService.cshtml
@@ -1,0 +1,67 @@
+@page "/Notifications/{NotificationId:int}/ShareWithService/{handler?}"
+
+@using ntbs_service.Helpers
+@using static NHSUK.FrontEndLibrary.TagHelpers.FormGroupType
+@model ntbs_service.Pages.AlertsAndActions.ShareWithServiceModel
+@using NHSUK.FrontEndLibrary.TagHelpers
+
+@{
+    Layout = "../Shared/_NotificationLayout.cshtml";
+    ViewData["Title"] = $"{Model.NotificationId} - Share with second service";
+}
+
+<nhs-width-container container-width="Standard">
+    <form method="post" autocomplete="off">
+        <h2> You are about to share this notification with a second TB service</h2>
+        <nhs-hint nhs-hint-type="Standard" id="ntbs-service-directory-hint">
+            You can search for TB services in the <a class="nhsuk-action-link" href=/ServiceDirectory>NTBS service directory</a> if you are not sure of the TB service name
+        </nhs-hint> <br>
+        <div>
+            @{
+                var hasTbServiceCodeError = !Model.ValidationService.IsValid("HospitalDetails.SecondaryTBServiceCode");
+                var tbServiceCodeGroupState = hasTbServiceCodeError ? Error : Standard;
+                var tbServiceCodeSelectErrorState = hasTbServiceCodeError ? SelectType.Error : SelectType.Standard;
+            }
+            <nhs-form-group nhs-form-group-type="@tbServiceCodeGroupState" id="HospitalDetails-SecondaryTBServiceCode" ref="tbServices">
+                <label nhs-label-type="Standard" asp-for="HospitalDetails.SecondaryTBServiceCode">
+                    @Html.DisplayNameFor(model => model.HospitalDetails.SecondaryTBServiceCode)
+                </label>
+                <span nhs-span-type="ErrorMessage" id="tb-service-error" has-error="@hasTbServiceCodeError"
+                      ref="errorField" asp-validation-for="HospitalDetails.SecondaryTBServiceCode"></span>
+                <autocomplete-select ref="filterContainer" placeholder="Select TB service" emit-change-on="select" inline-template>
+                    @* below id necessary to disable autocomplete in chrome - see https://adamsilver.io/blog/stopping-chrome-from-ignoring-autocomplete-off/ *@
+                    <select ref="selectField"
+                            nhs-select-type="@tbServiceCodeSelectErrorState"
+                            asp-for="HospitalDetails.SecondaryTBServiceCode"
+                            asp-items="Model.TbServices"
+                            id="selectTBService">
+                        <option value=""></option>
+                    </select>
+                </autocomplete-select>
+            </nhs-form-group>
+        </div>
+
+        @{
+            var hasReasonError = !Model.ValidationService.IsValid("HospitalDetails.ReasonForTBServiceShare");
+            var reasonFormGroupType = hasReasonError ? Error : Standard;
+        }
+
+        <nhs-form-group nhs-form-group-type="@reasonFormGroupType">
+            <label asp-for="HospitalDetails.ReasonForTBServiceShare" nhs-label-type="Standard">
+                Why are you sharing this record with another TB service?
+            </label>
+            <span id="reason-error" nhs-span-type="ErrorMessage" asp-validation-for="HospitalDetails.ReasonForTBServiceShare" has-error="@hasReasonError"></span>
+            <input asp-for="HospitalDetails.ReasonForTBServiceShare" nhs-input-type="Standard" is-error-input="@hasReasonError" aria-describedby="reason-error" />
+        </nhs-form-group>
+
+        <div class="flex-container">
+            <button id="confirm-share-button" nhs-button-type="Standard" classes="confirm-transfer-button ntbsuk-button--primary">
+                Confirm share
+            </button>
+
+            <nhs-back-link href="@RouteHelper.GetNotificationPath(Model.NotificationId, NotificationSubPaths.Overview)" classes="transfer-request-back-link">
+                Cancel & go back
+            </nhs-back-link>
+        </div>
+    </form>
+</nhs-width-container>
