+++ ntbs-service/Pages/AlertsAndActions/_CancelTransferConfirmation.cshtml
@@ -1,0 +1,16 @@
+@using ntbs_service.Helpers
+@model TransferRequestModel
+
+@{
+    Layout = "../Shared/_NotificationLayout.cshtml";
+    ViewData["Title"] = $"Transfer Request Cancelled - {Model.NotificationId}";
+}
+
+<nhs-width-container container-width="Standard">
+    <h2> Transfer of Notification #@Model.NotificationId has been cancelled </h2>
+    <br>
+    <a id="return-to-homepage" href="@RouteHelper.GetNotificationPath(Model.NotificationId, NotificationSubPaths.Overview)" 
+        class="nhsuk-button ntbsuk-button--primary return-to-homepage-link">
+        Return to the notification
+    </a>
+</nhs-width-container>
