+++ ntbs-service/Pages/Notifications/Denotify.cshtml.cs
@@ -42,7 +42,7 @@ namespace ntbs_service.Pages.Notifications
 
         public async Task<IActionResult> OnGetAsync()
         {
-            Notification = await NotificationRepository.GetNotificationAsync(NotificationId);
+            Notification = await GetNotificationAsync(NotificationId);
             if (Notification == null)
             {
                 return NotFound();
@@ -61,7 +61,7 @@ namespace ntbs_service.Pages.Notifications
 
         public async Task<IActionResult> OnPostConfirmAsync()
         {
-            Notification = await NotificationRepository.GetNotificationAsync(NotificationId);
+            Notification = await GetNotificationAsync(NotificationId);
 
             var (permissionLevel, _) = await _authorizationService.GetPermissionLevelAsync(User, Notification);
             if (permissionLevel != PermissionLevel.Edit)
