+++ ntbs-service/Pages/Notifications/LinkedNotifications.cshtml.cs
@@ -23,7 +23,7 @@ namespace ntbs_service.Pages.Notifications
 
         public async Task<IActionResult> OnGetAsync()
         {
-            Notification = await NotificationRepository.GetNotificationAsync(NotificationId);
+            Notification = await GetNotificationAsync(NotificationId);
             if (Notification == null)
             {
                 return NotFound();
