+++ ntbs-service/Pages/Notifications/NotificationChanges.cshtml.cs
@@ -31,7 +31,7 @@ namespace ntbs_service.Pages.Notifications
         // ReSharper disable once UnusedMember.Global
         public async Task<IActionResult> OnGetAsync()
         {
-            Notification = await NotificationRepository.GetNotificationAsync(NotificationId);
+            Notification = await GetNotificationAsync(NotificationId);
             if (Notification == null)
             {
                 return NotFound();
