+++ ntbs-service/Pages/Notifications/NotificationModelBase.cs
@@ -41,7 +41,6 @@ namespace ntbs_service.Pages.Notifications
         public DataQualityDraftAlert DraftAlert { get; set; }
         public PermissionLevel PermissionLevel { get; set; }
         public string PermissionReason { get; set; }
-        public bool IsShared { get; set; }
 
         [BindProperty(SupportsGet = true)]
         public int NotificationId { get; set; }
@@ -56,7 +55,6 @@ namespace ntbs_service.Pages.Notifications
             var (permissionLevel, reason) = await _authorizationService.GetPermissionLevelAsync(User, Notification);
             PermissionLevel = permissionLevel;
             PermissionReason = reason;
-            IsShared = PermissionLevel == PermissionLevel.SharedWith;
             NotificationBannerModel = new NotificationBannerModel(Notification, PermissionLevel == PermissionLevel.None);
         }
